+++ src/app/transactions/transactions.component.ts
@@ -85,7 +85,7 @@ export class TransactionsComponent implements OnInit {
     ]; 
 
     function balanceFormatter(params){
-      var stringbalance=parseInt(params.value);
+      var stringbalance=parseFloat(params.value);
       var twodecimalvalue=stringbalance.toFixed(2);
       return twodecimalvalue;
     }
